🔨 Move generic YAML types to `yaml.ts`
